src/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
bcrypt==4.2.0
APScheduler==3.10.4
mysqlclient==2.2.4
Flask-Bcrypt==1.0.1
Flask-APScheduler==1.13.1
